removing hard coded strategy, bug with underlying strategy, celery and logging
