+++ py-trading-bot/reporting/templates/reporting/reports.html
@@ -1,5 +1,5 @@
 <h1>Python Trading bot</h1>
-<p><b>Author: Psemdel, Version: 0.10, <a href="https://github.com/psemdel/py-trading-bot">Github</a></b></p>
+<p><b>Author: Psemdel, Version: 0.20, <a href="https://github.com/psemdel/py-trading-bot">Github</a></b></p>
 <table>
     <tr>
         <td><a href="{% url 'reporting:start_bot' %}">Start bot</a></td>
