+++ py-trading-bot/trading_bot/celery.py
@@ -4,6 +4,7 @@ from __future__ import absolute_import, unicode_literals
 
 import os
 from celery import Celery
+from celery.signals import setup_logging
 
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'trading_bot.settings')
@@ -19,6 +20,12 @@ app.config_from_object('django.conf:settings', namespace='CELERY')
 
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
+
+@setup_logging.connect
+def config_loggers(*args, **kwags):
+    from logging.config import dictConfig
+    from django.conf import settings
+    dictConfig(settings.LOGGING)
 
 @app.task(bind=True)
 def debug_task(self):
